+++ src/controllers/category.controller.ts
@@ -3,7 +3,7 @@ import { Response } from "express";
 import CategoryModel from "../models/category.model";
 import { categoryDAO } from "../models/category.model";
 import response from "../utils/response";
-import { object } from "yup";
+import { isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -61,6 +61,9 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId) {
+        return response.notFound(res, "category not found!");
+      }
       const result = await CategoryModel.findById(id);
 
       if (!result) {
@@ -76,6 +79,11 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId) {
+        return response.notFound(res, "failed update category");
+      }
+
       const result = await CategoryModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -89,6 +97,11 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId) {
+        return response.notFound(res, "failed remove category");
+      }
+
       const result = await CategoryModel.findByIdAndDelete(id, { new: true });
 
       response.success(res, result, "success update category");
