+++ src/controllers/ticket.controller.ts
@@ -59,6 +59,10 @@ export default {
     try {
       const { id } = req.params;
 
+      if (!isValidObjectId) {
+        return response.notFound(res, "ticket not found!");
+      }
+
       const result = await TicketModel.findById(id);
 
       if (!result) {
@@ -74,7 +78,9 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
-
+      if (!isValidObjectId) {
+        return response.notFound(res, "failed update ticket");
+      }
       const result = await TicketModel.findByIdAndUpdate(id, req.body, {
         new: true,
       });
@@ -87,6 +93,10 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId) {
+        return response.notFound(res, "failed remove ticket");
+      }
 
       const result = await TicketModel.findByIdAndDelete(id, { new: true });
       response.success(res, result, "success delete a ticket");
